updated layout for Xtend 2.9
this is mainly to add a UI subproject - however, the packaging has
also been improved.

In addition the troublesome antlr runtime has been included in the
feature. Should be possible now to install in a fairly vanilla eclipse
as long as the Xtext update site

http://download.eclipse.org/modeling/tmf/xtext/updates/composite/releases/

is included and searched for plugin resolution